Fixed bug for save query and visualizations
