+++ dashboards-observability/public/components/event_analytics/explorer/visualizations/index.tsx
@@ -37,7 +37,6 @@ interface IExplorerVisualizationsProps {
   visualizations: IVisualizationContainerProps;
   handleOverrideTimestamp: (field: IField) => void;
   callback?: any;
-  changeIsValidConfigOptionState: (isValidConfigOptionSelected: boolean) => void;
   queryManager: QueryManager;
 }
 
@@ -53,7 +52,6 @@ export const ExplorerVisualizations = ({
   visualizations,
   handleOverrideTimestamp,
   callback,
-  changeIsValidConfigOptionState,
   queryManager,
 }: IExplorerVisualizationsProps) => {
   const { vis } = visualizations;
@@ -156,7 +154,6 @@ export const ExplorerVisualizations = ({
                 curVisId={curVisId}
                 setCurVisId={setCurVisId}
                 callback={callback}
-                changeIsValidConfigOptionState={changeIsValidConfigOptionState}
               />
             </EuiResizablePanel>
           </>
